feat(kikan): V5b — move user + role entities + repos into kikan [#507]
Lifts platform auth primitives from `crates/core` + `crates/db` into
`crates/kikan::auth`:

- Domain types (UserId, RoleId, User, Role, CreateUser, UserRepository trait)
  from `mokumo_core::user` → `kikan::auth::domain`.
- SeaORM entities (users, roles) from `mokumo_db::{user,role}::entity`
  → `kikan::auth::{entity_user,entity_role}`.
- `SeaOrmUserRepo` + password helper from `mokumo_db::user::{repo,password}`
  → `kikan::auth::{repo,password}`.

`crates/db/src/{user,role}/mod.rs` retain thin re-export shims so existing
call sites (entity_drift tests, remaining services/api handlers slated
for V5c, BDD worlds) keep resolving against the old module paths until
`crates/db` dissolves in S3.1b.

`mokumo_core::user` is removed — callers now import from `kikan::auth`
directly. `crates/core` has no dependency on kikan, so a re-export
bridge from core to kikan is not available; the paths migrate in-place.

Repo tests move from inline `#[cfg(test)]` in `repo.rs` to
`crates/db/tests/user_repo.rs` as integration tests against the kikan
public surface. They need the full mokumo schema (users, roles seed,
settings, activity_log, updated_at triggers) which only the mokumo
migrator can produce; kikan cannot invoke that migrator without
violating I4. Tests relocate to the garment crate in S3.1b alongside
the rest of the vertical.

Quality gates green:
- cargo check --workspace --exclude mokumo-desktop --tests
- I1/I2/I4/I5 invariant checks
- kikan lib (77), mokumo-db lib (22), mokumo-db user_repo integration
  (21), mokumo-api lib (136) all pass
