+++ crates/core/src/lib.rs
@@ -7,4 +7,3 @@ pub mod pagination;
 pub mod sequence;
 pub mod setup;
 pub mod shop;
-pub mod user;
