+++ crates/kikan/src/auth/entity_role.rs
