+++ crates/kikan/src/auth/entity_user.rs
