+++ services/api/tests/auth_reset_regressions.rs
@@ -2,12 +2,12 @@ use std::path::PathBuf;
 
 use axum_test::TestServer;
 use kikan::SetupMode;
+use kikan::auth::SeaOrmUserRepo;
+use kikan::auth::UserRepository;
+use kikan::auth::{CreateUser, RoleId};
 use mokumo_api::auth::reset::recovery_file_path_for_email;
 use mokumo_api::{ServerConfig, build_app, ensure_data_dirs};
-use mokumo_core::user::traits::UserRepository;
-use mokumo_core::user::{CreateUser, RoleId};
 use mokumo_db::DatabaseConnection;
-use mokumo_db::user::repo::SeaOrmUserRepo;
 use serde_json::json;
 
 struct RunningServer {
